Refactor and reorganize components: removed unused files, updated imports, and improved styling
- Deleted ChatBoot, DataGridTable, Header, MessageItem, MessageList components
- Created new versions of ChatBoot, DataGridTable, Header, MessageItem, MessageList with updated structure and styles
- Adjusted styles for LineInputs and SelectLine components
- Enhanced responsiveness and visual consistency across components
- Updated HomePage to reflect new component paths
