+++ app/components/DataGridTable/DataGridTable.module.scss
@@ -1,0 +1,75 @@
+.dataGridTableContainer {
+  overflow-x: auto;
+  margin: 16px 0;
+}
+
+.dataGridTable {
+  width: 100%;
+  border-collapse: collapse;
+  background: #fff;
+  color: #111;
+  box-shadow: 0 2px 8px rgba(0,0,0,0.05);
+  border-radius: 8px;
+}
+
+.dataGridTable th,
+.dataGridTable td {
+  border: 1px solid #ccc;
+  padding: 8px;
+}
+
+.dataGridTable th {
+  background: #f3f3f3;
+  font-weight: bold;
+}
+
+.selectedRow {
+  background: #e3f2fd;
+}
+
+.selectedRowsContainer {
+  margin-top: 24px;
+  padding: 12px;
+  background: #f5f5f5;
+  border-radius: 6px;
+  border: 1px solid #ddd;
+}
+
+.selectedRowsTitle {
+  color: GrayText;
+  font-weight: bold;
+}
+
+.selectedRowsTable {
+  width: 100%;
+  border-collapse: collapse;
+  background: #fff;
+  color: #111;
+  margin-top: 8px;
+}
+
+.selectedRowsTable th,
+.selectedRowsTable td {
+  border: 1px solid #ccc;
+  padding: 8px;
+}
+
+.saveButton {
+  margin-top: 16px;
+  padding: 8px 24px;
+  background: #2196f3;
+  color: #fff;
+  border: none;
+  border-radius: 4px;
+  font-weight: bold;
+  cursor: pointer;
+}
+
+.successMessage {
+  margin-top: 12px;
+  color: #388e3c;
+  background: #e8f5e9;
+  padding: 8px 16px;
+  border-radius: 4px;
+  font-weight: bold;
+}
