+++ app/components/Header/Header.module.scss
@@ -1,0 +1,43 @@
+.header {
+  width: 100%;
+  background: #fff;
+  box-shadow: 0 2px 8px rgba(83, 181, 230, 0.08);
+  border-bottom: 1px solid #53b5e6;
+  border-radius: 12px;
+}
+
+.headerContent {
+  width: 100%;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  padding: 1rem;
+  border: 1px solid #53b5e6;
+}
+
+.logo {
+  display: flex;
+  align-items: center;
+}
+
+.title {
+  font-size: 1.85rem;
+  font-weight: bold;
+  color: #242424;
+  text-align: center;
+  border-radius: 8px;
+}
+
+@media (max-width: 768px) {
+  .headerContent {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+
+    .title {
+        margin-top: 0.8rem;
+      font-size: 1.5rem;
+    }
+  }
+}
