+++ app/components/LineInputs/LineInputs.module.scss
@@ -1,6 +1,7 @@
 .inputsAutomotiva {
   display: flex;
-  gap: 8px;
+  flex-wrap: wrap;
+  gap: 18px;
   margin-bottom: 16px;
 }
