+++ app/components/SelectLine/SelectLine.module.scss
@@ -10,7 +10,8 @@
   font-size: 16px;
   border-radius: 6px;
   border: 1px solid #bdbdbd;
-  background: #fff;
+  background: #bdbdbd;
+  color: #2c2b2b;
 }
 
 .buttonSubmit {
