+++ app/page.tsx
@@ -2,9 +2,9 @@
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 
-import Header from "./components/Header";
+import Header from "./components/Header/Header";
 
-import ChatBoot from "./components/ChatBoot";
+import ChatBoot from "./components/ChatBoot/ChatBoot";
 
 export default function HomePage() {
   return (
